+++ src/mouse_core/models/__init__.py
@@ -1,6 +1,6 @@
 from mouse_core.models.base import Model, load_model, save_model, push_model_to_hub
 from mouse_core.models.backbone import Backbone, LlamaBackbone, Qwen3Backbone, IdentityBackbone
-from mouse_core.models.heads import BaseHead, BaseHeadWithTarget, HeadSpec, SwiGLUHead, DiscreteActionValueHead, VectorActionValueHead, build_heads
+from mouse_core.models.heads import BaseHead, BaseHeadWithTarget, HeadSpec, SwiGLUHead, DiscreteActionValueHead, LayerwiseDiscreteActionValueHead, VectorActionValueHead, build_heads
 from mouse_core.models.embedding.embedding import Encoder, StepEmbedder, ModalitySpec
 
 __all__ = [
@@ -18,6 +18,7 @@ __all__ = [
     "HeadSpec",
     "SwiGLUHead",
     "DiscreteActionValueHead",
+    "LayerwiseDiscreteActionValueHead",
     "VectorActionValueHead",
     "build_heads",
     "ModalitySpec",